requirements.txt
watchdog==4.0.0
PyYAML==6.0.1
drain3==0.9.11
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
torch==2.2.1
fastapi==0.110.0
uvicorn==0.29.0
networkx==3.2.1
pgmpy==0.1.25
statsmodels==0.14.1
causal-learn==0.1.3.8
psycopg2-binary==2.9.9
jinja2==3.1.3
spacy==3.7.4
requests==2.31.0
matplotlib==3.8.3
seaborn==0.13.2
mlflow==2.11.1
dvc==3.48.4
transformers==4.38.2
h5py==3.10.0
redis==5.0.2
boto3==1.34.52
